Fit a Bayesian Rank-Clustered BTL model to ordinal comparison data (recommended for internal use only; use mcmc_RCBTL function instead)

This function fits a Bayesian Rank-Clustered BTL model to ordinal comparison data (e.g., complete rankings, partial rankings, pairwise comparisons, or groupwise comparisons) such that the worth parameters are drawn from a PSSF prior (as defined in Pearce and Erosheva 2024).

Usage

fit_RCBTL(
  Pi,
  J,
  a_gamma,
  b_gamma,
  lambda,
  nu0 = NULL,
  num_iters = 100,
  nu_reps = 3,
  groupwise = FALSE
)

Arguments

Pi

A matrix of rankings, such that the (i,j) entry is the index of the jth-most preferred object according to judge i. If groupwise=TRUE, then the index corresponds to the jth-most preferred object among those in row i; if groupwise=FALSE, it is assumd that all unranked objects (if any) are less preferred than those which are ranked.

J

A numeric indicating the total number of objects being compared.

a_gamma

A numeric for the first hyperparameter (shape) in a Gamma prior on each worth parameter.

b_gamma

A numeric for the second hyperparameter (rate) in a Gamma prior on each worth parameter.

lambda

A numeric for the Poisson hyperparameter on the number of non-empty clusters in the partition of worth parameters.

nu0

A numeric vector for the initialization of worth parameters, omega, in the MCMC algorithm. Default to NULL, indicating random initialization.

num_iters

A numeric indicating the total number of outer MCMC iterations (i.e., the number of times the partition is updated in the Gibbs sampler).

nu_reps

A numeric indicating the number of times each worth parameter is drawn per update of the parameter partition. There will be a total of num_itersxnu_reps samples from the posterior.

groupwise

A boolean to indicate whether the observed rankings are complete/partial rankings (FALSE; default) or groupwise comparisons (TRUE).

Value

A list of 4 elements: omega, a (num_itersxnu_reps)xJ matrix of approximate posterior draws of the object-specific worth parameters, omega; nu a (num_itersxnu_reps)xJ matrix of the unique parameter values corresponding to the jth partition cluster in posterior draw i, g a a (num_itersxnu_reps)xJ matrix indicating the cluster membership of object j in posterior draw i, and K a vector of the number of non-empty partition clusters in each posterior draw.

Examples

Pi <- matrix(data=c(1,2,3,NA,NA,1,2,3,4,5),byrow=TRUE,nrow=2)
fit_RCBTL(Pi=Pi,J=5,a_gamma=5,b_gamma=3,lambda=2,num_iters=5,nu_reps=2)
#> $omega
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#>  [1,] 1.2635258 0.6098669 1.2635258 0.6098669 1.2635258
#>  [2,] 1.4113309 1.1734638 1.4113309 1.1734638 1.4113309
#>  [3,] 1.2919746 1.0833011 1.2919746 1.0833011 1.2919746
#>  [4,] 1.9052428 1.4648513 1.9052428 1.4648513 1.9052428
#>  [5,] 1.6184554 3.4566871 1.6184554 3.4566871 1.6184554
#>  [6,] 1.9693112 1.1706151 1.9693112 1.1706151 1.9693112
#>  [7,] 1.4465826 0.7935678 1.4465826 0.7935678 1.4465826
#>  [8,] 1.0348992 0.5976839 1.0348992 0.5976839 1.0348992
#>  [9,] 0.9657842 4.0899460 0.9657842 1.6822816 0.9657842
#> [10,] 0.8758932 1.8412922 0.8758932 1.4020022 0.8758932
#> 
#> $nu
#>           [,1]      [,2]      [,3] [,4] [,5]
#>  [1,] 1.263526 0.6098669        NA   NA   NA
#>  [2,] 1.411331 1.1734638        NA   NA   NA
#>  [3,] 1.291975 1.0833011        NA   NA   NA
#>  [4,] 1.905243 1.4648513        NA   NA   NA
#>  [5,] 1.618455 3.4566871        NA   NA   NA
#>  [6,] 1.969311 1.1706151        NA   NA   NA
#>  [7,] 1.446583 0.7935678        NA   NA   NA
#>  [8,] 1.034899 0.5976839        NA   NA   NA
#>  [9,] 4.089946 1.6822816 0.9657842   NA   NA
#> [10,] 1.841292 1.4020022 0.8758932   NA   NA
#> 
#> $g
#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]    1    2    1    2    1
#>  [2,]    1    2    1    2    1
#>  [3,]    1    2    1    2    1
#>  [4,]    1    2    1    2    1
#>  [5,]    1    2    1    2    1
#>  [6,]    1    2    1    2    1
#>  [7,]    1    2    1    2    1
#>  [8,]    1    2    1    2    1
#>  [9,]    3    1    3    2    3
#> [10,]    3    1    3    2    3
#> 
#> $K
#>  [1] 2 2 2 2 2 2 2 2 3 3
#>